--- src/interface/VizBin/src/lcsb/vizbin/InvalidStateException.java
@@ -1,18 +1,0 @@
-package lcsb.vizbin;
-
-public class InvalidStateException extends RuntimeException {
-
-	public InvalidStateException(String string) {
-		super(string);
-	}
-
-	public InvalidStateException(Exception e) {
-		super(e);
-	}
-
-	/**
-	 * 
-	 */
-	private static final long	serialVersionUID	= 1L;
-
-}
