--- src/interface/VizBin/src/lcsb/vizbin/UnhandledOSException.java
@@ -1,14 +1,0 @@
-package lcsb.vizbin;
-
-public class UnhandledOSException extends Exception{
-
-	public UnhandledOSException(String string) {
-		super(string);
-	}
-
-	/**
-	 * 
-	 */
-	private static final long	serialVersionUID	= 1L;
-
-}
