--- src/interface/VizBin/src/lcsb/vizbin/data/Cluster.java
@@ -1,21 +1,0 @@
-package lcsb.vizbin.data;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class Cluster {
-	private List<Sequence>	elements	= new ArrayList<Sequence>();
-
-
-	public Cluster(List<Sequence> elements) {
-		this.elements = elements;
-	}
-
-	public List<Sequence> getElements() {
-		return elements;
-	}
-
-	public void setElements(ArrayList<Sequence> elements) {
-		this.elements = elements;
-	}
-}
