--- src/interface/VizBin/src/lcsb/vizbin/service/InvalidMetaFileException.java
@@ -1,13 +1,0 @@
-package lcsb.vizbin.service;
-
-public class InvalidMetaFileException extends Exception {
-
-	/**
-	 * 
-	 */
-	private static final long	serialVersionUID	= 1L;
-
-	public InvalidMetaFileException(String text) {
-		super(text);
-	}
-}
