+++ src/interface/VizBin/src/lu/uni/lcsb/vizbin/UnhandledOSException.java
@@ -1,0 +1,14 @@
+package lu.uni.lcsb.vizbin;
+
+public class UnhandledOSException extends Exception{
+
+	public UnhandledOSException(String string) {
+		super(string);
+	}
+
+	/**
+	 * 
+	 */
+	private static final long	serialVersionUID	= 1L;
+
+}
