+++ src/interface/VizBin/src/lu/uni/lcsb/vizbin/service/InvalidMetaFileException.java
@@ -1,0 +1,13 @@
+package lu.uni.lcsb.vizbin.service;
+
+public class InvalidMetaFileException extends Exception {
+
+	/**
+	 * 
+	 */
+	private static final long	serialVersionUID	= 1L;
+
+	public InvalidMetaFileException(String text) {
+		super(text);
+	}
+}
