+++ src/interface/VizBin/src/lu/uni/lcsb/vizbin/service/utils/PcaType.java
@@ -1,0 +1,15 @@
+package lu.uni.lcsb.vizbin.service.utils;
+
+public enum PcaType {
+	MTJ("Mtj"), MTJ_OPTIMIZED("Mtj optimized"), EJML("EJML");
+
+	String	commonName	= null;
+
+	private PcaType(String name) {
+		commonName = name;
+	}
+
+	public String getName() {
+		return commonName;
+	}
+}
