--- src/interface/VizBin/test/lcsb/vizbin/service/utils/pca/AllPcaTests.java
@@ -1,11 +1,0 @@
-package lcsb.vizbin.service.utils.pca;
-
-import org.junit.runner.RunWith;
-import org.junit.runners.Suite;
-import org.junit.runners.Suite.SuiteClasses;
-
-@RunWith(Suite.class)
-@SuiteClasses({ PrincipleComponentAnalysisEJMLTest.class, PrincipleComponentAnalysisMtjTest.class })
-public class AllPcaTests {
-
-}
